+++ src/main/java/com/zagorskidev/carpricing/service/regression/RegressionService.java
@@ -1,10 +1,8 @@
 package com.zagorskidev.carpricing.service.regression;
 
-import java.math.BigDecimal;
-
 import com.zagorskidev.carpricing.domain.CarParameters;
 
 public interface RegressionService 
 {
-	public BigDecimal predict(CarParameters params);
+	public Double predict(CarParameters params);
 }
